# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directory of Organizations API",
        "description": "API для работы со справочником организаций, зданий и видов деятельности",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/activity": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить список всех видов деятельности",
                "operationId": "f0f4b5861cc9d5197d4ef619359ab1a1",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Создать новый вид деятельности",
                "operationId": "ef70eaa884bb37ab50a64c9f61d3753d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вид деятельности создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/activity/tree": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить древовидную структуру видов деятельности",
                "operationId": "77e1b8060bab6ec19aacc6207bec0bd6",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/activity/{activityId}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить вид деятельности по ID",
                "operationId": "8ed3355eee27cbefa5c33539a3dee22b",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Удалить вид деятельности",
                "operationId": "022a241c7c5a3c7a936d8ff9478fa862",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вид деятельности удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Activity by id: 1 deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Activities"
                ],
                "summary": "Обновить вид деятельности",
                "operationId": "84b186347e591d6c529477a6c65ef725",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вид деятельности обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/activity/by-activity/{activity}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Поиск компаний по виду деятельности (включая дочерние)",
                "operationId": "0492a7532238a47fa8e4b4ba257f3c62",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "path",
                        "description": "Идентификатор вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "activity": {
                                            "type": "string"
                                        },
                                        "companies_count": {
                                            "type": "integer"
                                        },
                                        "companies": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Activity not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "IT-услуги"
                    },
                    "identifier": {
                        "type": "string",
                        "example": "it_services"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
            "ActivityStoreRequest": {
                "required": [
                    "name",
                    "identifier"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "IT-услуги"
                    },
                    "identifier": {
                        "type": "string",
                        "example": "it_services"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ActivityUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Обновленное название"
                    },
                    "identifier": {
                        "type": "string",
                        "example": "updated_identifier"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Company": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Ромашка"
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Static API key authentication",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}